Made LightningAnalyzer template class and based Lightcone off it
